obstetric-ai-system/requirements-dev.txt
# Développement local - tous agents
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
httpx>=0.26.0
anthropic>=0.18.0
numpy>=1.24.0
